requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
xgboost
flask
python-dotenv
certifi
pyaml
mlflow
pymongo
dagshub
uvicorn
fastapi
pymongo[srv]==3.12